requirements.txt
# Deep Learning
torch>=2.4.0
transformers>=4.49.0
accelerate>=1.0.0
deepspeed>=0.15.0
peft>=0.14.0

# Training Framework — TRL (unified SFT / DPO / PPO interface for transformers)
trl>=0.15.0

# Vision & Multimodal
qwen-vl-utils>=0.0.10
pillow>=10.0.0

# Evaluation
rouge-score>=0.1.2
scikit-learn>=1.5.0
jieba>=0.42.1
nltk>=3.9.0

# API & Functional Evaluation
openai>=1.50.0
httpx>=0.27.0

# Utilities
tqdm>=4.66.0
numpy>=1.26.0
pandas>=2.2.0
pyyaml>=6.0.2
jsonlines>=4.0.0

# Optional: for Google Drive weight download
gdown>=5.2.0
